+++ test/filter-cloudwatch-compiler.spec.js
@@ -18,6 +18,7 @@ function verify_compile_error(source, filter) {
 
     try {
         compiler.compile(ast);
+        throw new Error('Compile succeeded when it should have failed');
     } catch (e) {
         expect(e).to.be.instanceOf(JuttleErrors.CompileError);
         expect(e.code).to.equal('RT-ADAPTER-UNSUPPORTED-FILTER');
@@ -41,7 +42,7 @@ function verify_compile_success(source, expected) {
 
 describe('aws filter', function() {
 
-    describe(' properly returns errors for invalid filtering expressions like ', function() {
+    describe('properly returns errors for invalid filtering expressions like', function() {
 
         var invalid_unary_operators = ['!', '-'];
 
@@ -60,9 +61,79 @@ describe('aws filter', function() {
             });
         });
 
-        it('Combining terms with AND', function() {
+        it('Combining products with AND', function() {
             verify_compile_error('product="EC2" AND product="EBS"',
-                                 'operator AND');
+                                 'AND between products');
+        });
+
+        it('Combining items with AND', function() {
+            verify_compile_error('product="EC2" AND item="i-cb955911" AND item="i-966a694d"',
+                                 'AND between items');
+        });
+
+        it('Combining items with AND (using concise item notation)', function() {
+            verify_compile_error('item="EC2:i-cb955911" AND item="EC2:i-966a694d"',
+                                 'AND between products');
+        });
+
+        it('Combining product and item with AND', function() {
+            verify_compile_error('product="EC2" AND item="EC2:i-cb955911"',
+                                 'AND between products');
+        });
+
+        it('Combining product and item with AND (different products)', function() {
+            verify_compile_error('product="EC2" AND item="EBS:vol-56130db1"',
+                                 'AND between products');
+        });
+
+        it('Combining product and metric with AND', function() {
+            verify_compile_error('product="EC2" AND metric="EC2:DiskReadBytes"',
+                                 'AND between products');
+        });
+
+        it('Combining product and metric with AND (different products)', function() {
+            verify_compile_error('product="EC2" AND metric="EBS:DiskReadBytes"',
+                                 'AND between products');
+        });
+
+        it('Combining product, metric and item with AND (different products)', function() {
+            verify_compile_error('product="EBS" AND metric="DiskReadBytes" and item="EC2:i-cb955911"',
+                                 'AND between products');
+        });
+
+        it('Combining metrics with AND', function() {
+            verify_compile_error('product="EC2" AND metric="CPUUtilization" AND metric="DiskReadOps"',
+                                 'AND between metrics');
+        });
+
+        it('Combining metrics with AND (concise notation)', function() {
+            verify_compile_error('metric="EC2:CPUUtilization" AND metric="EC2:DiskReadOps"',
+                                 'AND between products');
+        });
+
+        it('Combining groups of ORs with AND (on one side)', function() {
+            verify_compile_error('(product="EC2" OR product="EBS") AND product="RDS"',
+                                 'AND between anything other than simple conditions');
+        });
+
+        it('Combining groups of ORs with AND (on both sides)', function() {
+            verify_compile_error('(product="EC2" OR product="EBS") AND (product="RDS" OR product="EC2")',
+                                 'AND between anything other than simple conditions');
+        });
+
+        it('Nested ANDs and ORs)', function() {
+            verify_compile_error('(metric="DiskReadBytes" OR metric="DiskWriteBytes") AND (metric="NetworkIn" OR metric="NetworkOut")',
+                                 'AND between anything other than simple conditions');
+        });
+
+        it('Combining groups of item ORs with AND)', function() {
+            verify_compile_error('(item="i-cb955911" OR item="i-11cb9559") AND (item="i-cc696a17" OR item="i-17cc696a")',
+                                 'AND between anything other than simple conditions');
+        });
+
+        it('Combining nested groups of ORs with AND', function() {
+            verify_compile_error('((product="EC2" AND item="i-cb955911") OR (product="EC2" and item="i-11cb9559")) AND ((product="RDS" AND item="db-production") OR (product="EBS" and metric="DiskWriteBytes"))',
+                                 'AND between anything other than simple conditions');
         });
 
         it('Using NOT on a term', function() {
@@ -76,13 +147,18 @@ describe('aws filter', function() {
         });
 
         it('matching on unsupported products', function() {
-            verify_compile_error('product = "RDS"',
-                                 'product RDS');
+            verify_compile_error('product = "Lambda"',
+                                 'product Lambda');
         });
 
-        it('item in filter not having format <aws product>:<item name>', function() {
+        it('item in filter without any corresponding product', function() {
             verify_compile_error('item = "i-cb955911"',
-                                 'item value not having format <aws product>:<item name>');
+                                 'item/metric condition without product');
+        });
+
+        it('metric in filter without any corresponding product', function() {
+            verify_compile_error('metric = "DiskReadBytes"',
+                                 'item/metric condition without product');
         });
 
         it('item in filter for unsupported product', function() {
@@ -92,66 +168,343 @@ describe('aws filter', function() {
 
         it('a single filter expression', function() {
             verify_compile_error('\"foo\"',
-                                 'filter term UnaryExpression');
+                                 'simple filter terms');
         });
 
         it('not a filter expression or string', function() {
             verify_compile_error('+ 1',
-                                 'operator +');
+                                 'simple filter terms');
         });
     });
 
-    describe(' properly returns condition lists for valid cases like ', function() {
+    describe('properly returns condition lists for valid cases like', function() {
         it('Single product match', function() {
-            verify_compile_success('product="EC2"', {
-                EC2: []
-            });
+            verify_compile_success('product="EC2"', [{
+                product: 'EC2',
+                item: [],
+                metric: []
+            }]);
         });
 
         it('Multiple product matches', function() {
-            verify_compile_success('product="EC2" OR product="EBS"', {
-                EC2: [],
-                EBS: []
-            });
+            verify_compile_success('product="EC2" OR product="EBS"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: []
+                }
+            ]);
         });
 
         it('Multiple product matches w/ duplicates', function() {
-            verify_compile_success('product="EC2" OR product="EC2"', {
-                EC2: []
-            });
+            verify_compile_success('product="EC2" OR product="EC2"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                },
+            ]);
+        });
+
+        it('Multiple product matches, not adjacent', function() {
+            verify_compile_success('product="EC2" OR product="EBS" OR product="EC2"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: []
+                },
+            ]);
         });
 
         it('Single item match', function() {
-            verify_compile_success('item="EC2:i-cc696a17"', {
-                EC2: ['i-cc696a17']
-            });
+            verify_compile_success('item="EC2:i-cc696a17"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cc696a17'],
+                    metric: []
+                }
+            ]);
         });
 
         it('Multiple item matches', function() {
-            verify_compile_success('item="EC2:i-cc696a17" OR item="EBS:vol-56130db1"', {
-                EBS: ['vol-56130db1'],
-                EC2: ['i-cc696a17']
-            });
+            verify_compile_success('item="EC2:i-cc696a17" OR item="EBS:vol-56130db1"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cc696a17'],
+                    metric: []
+                },
+                {
+                    product: 'EBS',
+                    item: ['vol-56130db1'],
+                    metric: []
+                }
+            ]);
         });
 
         it('Multiple item matches for same product', function() {
-            verify_compile_success('item="EC2:i-cc696a17" OR item="EC2:i-966a694d"', {
-                EC2: ['i-cc696a17', 'i-966a694d']
-            });
+            verify_compile_success('item="EC2:i-cc696a17" OR item="EC2:i-966a694d"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cc696a17', 'i-966a694d'],
+                    metric: []
+                }
+            ]);
         });
 
         it('Multiple item matches with duplicates', function() {
-            verify_compile_success('item="EC2:i-cc696a17" OR item="EC2:i-cc696a17"', {
-                EC2: ['i-cc696a17']
-            });
+            verify_compile_success('item="EC2:i-cc696a17" OR item="EC2:i-cc696a17"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cc696a17'],
+                    metric: []
+                }
+            ]);
+        });
+
+        it('Single metric match', function() {
+            verify_compile_success('metric="EC2:CPUUtilization"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['CPUUtilization']
+                }
+            ]);
+        });
+
+        it('Multiple metric matches', function() {
+            verify_compile_success('metric="EC2:CPUUtilization" OR metric="EBS:VolumeReadBytes"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['CPUUtilization']
+                },
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: ['VolumeReadBytes']
+                }
+            ]);
+
+        });
+
+        it('Multiple metric matches for same product', function() {
+            verify_compile_success('metric="EC2:CPUUtilization" OR metric="EC2:DiskReadOps"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['CPUUtilization', 'DiskReadOps']
+                }
+            ]);
+        });
+
+        it('Multiple item matches with duplicates', function() {
+            verify_compile_success('metric="EC2:CPUUtilization" OR metric="EC2:CPUUtilization"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['CPUUtilization']
+                }
+            ]);
+        });
+
+        it('Combining product and metrics with AND', function() {
+            verify_compile_success('product="EC2" AND metric="DiskReadOps"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['DiskReadOps']
+                }
+            ]);
+        });
+
+        it('Combining products, metrics, and items with AND', function() {
+            verify_compile_success('product="EC2" AND item="i-cb955911" AND metric="DiskReadOps"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cb955911'],
+                    metric: ['DiskReadOps']
+                }
+            ]);
+        });
+
+        it('Combining products, metrics, and items with AND (concise notation)', function() {
+            verify_compile_success('item="EC2:i-cb955911" AND metric="DiskReadOps"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cb955911'],
+                    metric: ['DiskReadOps']
+                }
+            ]);
+        });
+
+        it('Combining products, metrics, and items with AND, along with unrelated products', function() {
+            verify_compile_success('item="EC2:i-cb955911" AND metric="DiskReadOps" OR product="EBS" OR product="RDS"', [
+                {
+                    product: 'EC2',
+                    item: ['i-cb955911'],
+                    metric: ['DiskReadOps']
+                },
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'RDS',
+                    item: [],
+                    metric: []
+                }
+            ]);
+        });
+
+        it('Combining groups of products/metrics/items combined with AND', function() {
+            verify_compile_success('(product="EC2" AND item="i-cb955911" AND metric="DiskReadOps") OR (product="EBS" and metric="DiskWriteBytes") OR (product="RDS" and item="db-production")', [
+                {
+                    product: 'EC2',
+                    item: ['i-cb955911'],
+                    metric: ['DiskReadOps']
+                },
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: ['DiskWriteBytes']
+                },
+                {
+                    product: 'RDS',
+                    item: ['db-production'],
+                    metric: []
+                }
+            ]);
+        });
+
+        it('Separate product items can be merged', function() {
+            verify_compile_success('(product="EC2" AND item="i-cb955911") OR product="RDS" OR (product="EC2" AND item="i-cc696a17")', [
+                {
+                    product: 'EC2',
+                    item: ['i-cb955911', 'i-cc696a17'],
+                    metric: []
+                },
+                {
+                    product: 'RDS',
+                    item: [],
+                    metric: []
+                }
+            ]);
+        });
+
+        it('Single and wildcard product items will not be merged', function() {
+            verify_compile_success('product="EC2" OR product="RDS" OR (product="EC2" AND item="i-cc696a17")', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'RDS',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EC2',
+                    item: ['i-cc696a17'],
+                    metric: []
+                },
+            ]);
+        });
+
+        it('Single and wildcard product metrics will not be merged', function() {
+            verify_compile_success('product="EC2" OR product="RDS" OR (product="EC2" AND metric="CPUUtilization")', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'RDS',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['CPUUtilization']
+                },
+            ]);
+        });
+
+        it('Separate product metrics can be merged', function() {
+            verify_compile_success('(product="EBS" AND metric="DiskReadBytes") OR product="EC2" OR (product="EBS" AND metric="DiskWriteBytes")', [
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: ['DiskReadBytes', 'DiskWriteBytes']
+                },
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                }
+            ]);
+        });
+
+        it('Separate product items/metric conditions will not be merged', function() {
+            verify_compile_success('(product="EC2" AND item="i-cb955911") OR product="RDS" OR (product="EC2" AND metric="CPUUtilization")', [
+                {
+                    product: 'EC2',
+                    item: ['i-cb955911'],
+                    metric: []
+                },
+                {
+                    product: 'RDS',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: ['CPUUtilization']
+                }
+            ]);
         });
 
         it('Mix of item and product matches', function() {
-            verify_compile_success('product="EC2" OR product="EBS" OR item="EC2:i-cc696a17" OR item="EC2:i-966a694d" OR item="EBS:vol-56130db1" OR product="RDS"', {
-                EBS: ['vol-56130db1'],
-                EC2: ['i-cc696a17', 'i-966a694d'],
-                RDS: []
-            });
+            verify_compile_success('product="EC2" OR product="EBS" OR item="EC2:i-cc696a17" OR item="EC2:i-966a694d" OR item="EBS:vol-56130db1" OR product="RDS"', [
+                {
+                    product: 'EC2',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EBS',
+                    item: [],
+                    metric: []
+                },
+                {
+                    product: 'EC2',
+                    item: ['i-cc696a17', 'i-966a694d'],
+                    metric: []
+                },
+                {
+                    product: 'EBS',
+                    item: ['vol-56130db1'],
+                    metric: []
+                },
+                {
+                    product: 'RDS',
+                    item: [],
+                    metric: []
+                }
+            ]);
         });
 
     });